Viewmodel

+++ Actor/LaserEmitter.c
@@ -47,7 +47,7 @@ bool LaserEmitterSignalHandler(Actor *this, const Actor *sender, byte signal, co
 	if (signal == LASER_EMITTER_INPUT_ENABLE)
 	{
 		ActorTriggerInput(this, data->laserActor, LASER_INPUT_ENABLE, NULL);
-		this->currentSkinIndex = (int)data->height + 1;
+		this->currentSkinIndex = data->height + 1;
 		return true;
 	}
 	return false;
@@ -77,7 +77,7 @@ void LaserEmitterInit(Actor *this, b2WorldId worldId, const KvList *params)
 	data->hasTicked = false;
 
 	this->actorModel = LoadModel(MODEL("model_laseremitter"));
-	this->currentSkinIndex = (int)data->height + 1;
+	this->currentSkinIndex = data->height + 1;
 
 	CreateLaserEmitterCollider(this, worldId);
